refactor autos widgets
